requirements.txt
replicate==0.24.0
python-dotenv==1.0.1
boto3==1.34.59
flask==3.0.2